requirements.txt
streamlit
opencv-python-headless
onnxruntime
numpy
Pillow